@@ -10,21 +10,20 @@ authors = [{ name = "Janosh Riebesell", email = "janosh.riebesell@gmail.com" }]
 license = { file = "license" }
 readme = 'readme.md'
 keywords = [
-    "interactive",
-    "python",
-    "jupyter",
-    "ipython",
     "code formatting",
+    "interactive",
+    "ipython",
+    "jupyter",
     "pre-commit hook",
+    "python",
 ]
 classifiers = [
-    "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
     "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3 :: Only",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 
 [project.urls]
 Homepage = "https://github.com/janosh/format-ipy-cells"
@@ -52,30 +51,17 @@ warn_unused_ignores = true
 no_implicit_optional = false
 
 [tool.ruff]
-target-version = "py38"
-select = [
-    "B",   # flake8-bugbear
-    "D",   # pydocstyle
-    "E",   # pycodestyle
-    "ERA", # flake8-eradicate
-    "F",   # pyflakes
-    "I",   # isort
-    "PLE", # pylint error
-    "PLW", # pylint warning
-    "PYI", # flakes8-pyi
-    "Q",   # flake8-quotes
-    "SIM", # flake8-simplify
-    "TID", # tidy imports
-    "UP",  # pyupgrade
-    "W",   # pycodestyle
-    "YTT", # flake8-2020
-]
+target-version = "py310"
+select = ["ALL"]
 ignore = [
-    "D100", # Missing docstring in public module
-    "D104", # Missing docstring in public package
-    "D205", # 1 blank line required between summary line and description
+    "D100",   # Missing docstring in public module
+    "D104",   # Missing docstring in public package
+    "D205",   # 1 blank line required between summary line and description
+    "INP001", # implicit-namespace-packages
+    "PTH",    # prefer pathlib over os.path
+    "T201",   # print
 ]
 pydocstyle.convention = "google"
 
 [tool.ruff.per-file-ignores]
-"tests/*" = ["D103"]
+"tests/*" = ["D103", "S101"]
